requirements.txt
streamlit==1.28.1
langchain-groq==0.1.9
langchain-core==0.2.38
pymongo==4.6.0
python-dotenv==1.0.0
